// frontend/src/shared/types/enums.ts
export enum EstadoPedido {
  PENDIENTE = 'pendiente',
  EN_AGUA = 'en_agua',
  FINALIZADO = 'finalizado',
  CANCELADO = 'cancelado',
}

export enum TipoMovimiento {
  ENTRADA = 'entrada',
  SALIDA = 'salida',
  MOVIMIENTO_CUNA = 'movimiento_cuna',
}

export enum EstadoEmbarcacion {
  EN_CUNA = 'EN_CUNA',
  EN_AGUA = 'EN_AGUA',
  EN_MANTENIMIENTO = 'EN_MANTENIMIENTO',
  INACTIVA = 'INACTIVA',
  BLOQUEADA = 'BLOQUEADA',
}

export enum EstadoSolicitud {
  PENDIENTE = 'PENDIENTE',
  EN_AGUA = 'EN_AGUA',
  FINALIZADA = 'FINALIZADA',
  CANCELADA = 'CANCELADA',
}